// src/module/lote/enums/tipo-evento-trazabilidad.enum.ts
export enum TipoEventoTrazabilidad {
  RECEPCION = 'RECEPCION',
  CLASIFICACION = 'CLASIFICACION',
  REVISION_CALIDAD = 'REVISION_CALIDAD',
  CAMBIO_UBICACION = 'CAMBIO_UBICACION',
  INGRESO_CAMARA = 'INGRESO_CAMARA',
  CONSUMO_PARCIAL = 'CONSUMO_PARCIAL',
  FINALIZACION = 'FINALIZACION',
  RECOMENDACION_DESTINO = 'RECOMENDACION_DESTINO', 
}